fix(console): finish type mapping (mysql-adbmysql.yml、mysql-oracle.yml、mysql-postgres.yml)

+++ src/test/resources/net_hasor_db/all_types/tb_adb4mysql_types.sql
@@ -56,9 +56,8 @@ create table tb_adbmysql_types
     c_float_z       float zerofill,
     c_float_n       float(10),
     c_float_n_z     float(10) zerofill,
-    c_float_n_n     float(10, 3
-) ,
-    c_float_n_n_z            float(10, 3) zerofill,
+    c_float_n_n     float(10, 3) ,
+    c_float_n_n_z   float(10, 3) zerofill,
 
     c_double                 double,
     c_double_z               double zerofill,
